introduce resources: getDataResource and getStaticResource
Resources are used in lieu of fetching from the data dir.
In normal builds, they *do* fetch from the data dir.
In production builds, those files are embedded into the executable itself

Static file serving had to change to use these, and so there is now a minimal staticApp (in Server.hs)
This entailed having our own compiled in mime table (should be read from a file one day)
Hence, there is no longer a dependency on the package wai-middleware-static