FD clean up
Ensure FDs created in the Job machinery are properly closed in forked midground processes.
This means ensuring that FDs only for the master shell thread are closed in each child, and that
various other descriptors are set "close on exec".

Added a closeFdSafe, as sometimes the original standard out FD is closed already (not sure why!)