Change default build flag setting so all-in-one executable is built by default
The build flag changed from Production to live-files.
The default for cabal configure is for live-files to be false.
Improved the README to match both new build default, and to better explain using the --local and --remote options
Impoved the --help display a small bit w.r.t. --local and --remote options.