improve autoscrolling output by simplifying it
now, if the output is scrolled to the bottom, then rescroll to attempt to show the last chunk of output
if the chunk is taller than the viewport, then the top portion will be in view - and auto-scrolling will
top because the output isn't scrolled to the bottom. If it all fits, then auto-scroling will continue. As a
bonus, if the user scrolls during output, it will stop auto-scrolling until the user gets to the bottom.
Seems to work reasonably...